Crate namada_events

Events emitted by the Namada ledger.

Modules

extend

Extend events with additional fields.

Macros

event_type

Instantiate a new EventType in const contexts. Mostly useful to define new event types in the protocol.

Structs

Event

Custom events that can be queried from Tendermint using a websocket client

EventType

ABCI event type.

EventTypeBuilder

Build an EventType segment by segment.

Enums

EventError

Errors to do with emitting events.

EventLevel

Indicates if an event is emitted do to an individual Tx or the nature of a finalized block

Traits

EmitEvents

Used in sub-systems that may emit events.

EventToEmit

An event to be emitted in Namada.