namada_events
Module extend

Extend events with additional fields.

Structs

Closure

Extend an Event with the given closure.

CompositeEvent

Event composed of various other event extensions.

Domain

Extend an Event with a new domain.

EventValue

Implement the Display and FromStr traits for any serde type

Height

Extend an Event with block height information.

Info

Extend an Event with info data.

InnerTxHash

Extend an Event with inner transaction hash information.

Log

Extend an Event with log data.

ParsedDomain

Parsed domain of some event.

TxHash

Extend an Event with transaction hash information.

Enums

UserAccount

A user account.

Traits

AttributesMap

Map of event attributes.

ComposeEvent

Provides event composition routines.

DeleteFromEventAttributes

Delete an attribute from an event’s attributes.

EventAttributeChecker

Checks for the presence of an attribute in the provided attributes map.

EventAttributeEntry

Represents an entry in the attributes of an Event.

ExtendAttributesMap

Extend an AttributesMap implementation with the ability to add new attributes from domain types.

ExtendEvent

Extend an event with additional data.

ExtendEventAttributes

Extend an event with additional attributes.

RawReadFromEventAttributes

Read a raw (string encoded) attribute from an event’s attributes.

ReadFromEventAttributes

Read an attribute from an event’s attributes.

Functions

attribute_checker

Return a new implementation of EventAttributeChecker.

dispatch_attribute

Dispatch a callback on a list of attribute kinds.

event_domain_of

Build a new Domain to extend an event with.