namada_events::extend
Trait EventAttributeChecker

pub trait EventAttributeChecker<'value, A>
where
    A: AttributesMap,
{
    // Required method
    fn is_present(&self, attributes: &A) -> bool
       where A: AttributesMap;
}

Checks for the presence of an attribute in the provided attributes map.

Required Methods

fn is_present(&self, attributes: &A) -> bool
where A: AttributesMap,

Check if the associated attribute is present in the provided event attributes.