namada_events::extend
Trait RawReadFromEventAttributes

pub trait RawReadFromEventAttributes<'value> {
    // Required methods
    fn check_if_present_in<A>(attributes: &A) -> bool
       where A: AttributesMap;
    fn raw_read_opt_from_event_attributes<A>(attributes: &A) -> Option<&str>
       where A: AttributesMap;
    fn raw_read_from_event_attributes<A>(
        attributes: &A,
    ) -> Result<&str, EventError>
       where A: AttributesMap;
}

Read a raw (string encoded) attribute from an event’s attributes.

Required Methods

fn check_if_present_in<A>(attributes: &A) -> bool
where A: AttributesMap,

Check if the associated attribute is present in the provided event attributes.

fn raw_read_opt_from_event_attributes<A>(attributes: &A) -> Option<&str>
where A: AttributesMap,

Read a string encoded attribute from the provided event attributes.

fn raw_read_from_event_attributes<A>(attributes: &A) -> Result<&str, EventError>
where A: AttributesMap,

Read a string encoded attribute from the provided event attributes.

Object Safety

This trait is not object safe.